# api/openapi/spec.yaml
openapi: 3.1.0
info:
  title: Rncasp API
  description: >
    REST API for the Rncasp (Really No Clue About Shift Planning) application.
    A shift planning web application for LAN party events that replaces Excel-based
    workflows where team members sign up for shifts across multi-day events.


    All responses (except health and binary downloads) are wrapped in a standard
    envelope: `{"data": ..., "error": ..., "meta": ...}`.


    Authentication is cookie-based using session cookies set on login/register.
  version: 1.0.0
  contact:
    name: Rncasp
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
...
security:
  - cookieAuth: []
paths:
...
  /api/users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List users
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [super_admin, user, read_only]
        - name: account_type
          in: query
          schema:
            type: string
            enum: [local, oauth, dummy]
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/users/search:
    get:
      tags: [Users]
      operationId: searchUsers
      summary: Search users by query string
      parameters:
        - name: q
          in: query
          description: Search query (matches username, full_name, display_name)
          schema:
            type: string
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/users/{userId}:
    get:
      tags: [Users]
      operationId: getUser
      summary: Get a user by ID
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Users]
      operationId: updateUser
      summary: Update a user
      description: >
        Super-admin only. Can change role, active status, account type,
        name, email, password, and time format.
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/users/dummy:
    post:
      tags: [Users]
      operationId: createDummy
      summary: Create a dummy account
      description: >
        Super-admin only. Dummy accounts are placeholder users that
        cannot log in but can be assigned to shifts by admins.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDummyRequest"
      responses:
        "201":
          description: Dummy account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/users/dummy/{userId}:
    put:
      tags: [Users]
      operationId: updateDummy
      summary: Update a dummy account
      description: Super-admin only.
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDummyRequest"
      responses:
        "200":
          description: Dummy account updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Users]
      operationId: deleteDummy
      summary: Delete a dummy account
      description: Super-admin only.
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Dummy account deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # iCal Tokens
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: >
        Session cookie set by the login or register endpoints.
        The cookie name is configurable; `session` is the default.

  # ---------------------------------------------------------------------------
  # Parameters
  # ---------------------------------------------------------------------------
  parameters:
...
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Limit:
      name: limit
      in: query
      description: Maximum number of results to return
      schema:
        type: integer
        default: 50
        minimum: 1
    Offset:
      name: offset
      in: query
      description: Number of results to skip
      schema:
        type: integer
        default: 0
        minimum: 0

  # ---------------------------------------------------------------------------
  # Responses
  # ---------------------------------------------------------------------------
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: invalid_input
              message: "invalid request body"
    Unauthorized:
      description: Not authenticated or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: unauthorized
              message: "not authenticated"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: forbidden
              message: "insufficient permissions"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: not_found
              message: "resource not found"
    Conflict:
      description: Resource already exists or conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: already_exists
              message: "resource already exists"
...
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/APIError"
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          enum:
            - not_found
            - already_exists
            - conflict
            - unauthorized
            - forbidden
            - invalid_input
            - event_locked
            - overbooking_not_allowed
            - account_inactive
            - dummy_account_cannot_login
            - invalid_totp_code
            - totp_required
            - too_many_requests
        message:
          type: string
          description: Human-readable error message
        field:
          type: string
          description: Field name for validation errors
...
    MessageResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            message:
              type: string
          required: [message]

    # --- Health ---
...
    User:
      type: object
      required:
        - id
        - username
        - full_name
        - role
        - language
        - account_type
        - time_format
        - totp_enabled
        - is_active
        - created_at
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        full_name:
          type: string
        display_name:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        role:
          type: string
          enum: [super_admin, user, read_only]
        language:
          type: string
          enum: [en, de]
        account_type:
          type: string
          enum: [local, oauth, dummy]
        time_format:
          type: string
          enum: ["24h", "12h"]
        totp_enabled:
          type: boolean
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    UserResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/User"

    # --- Auth requests ---
...
    UpdateUserRequest:
      type: object
      properties:
        role:
          type: string
          enum: [super_admin, user, read_only]
        is_active:
          type: boolean
        full_name:
          type: string
        display_name:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        password:
          type: string
        time_format:
          type: string
          enum: ["24h", "12h"]
        username:
          type: string
        account_type:
          type: string
          enum: [local, oauth, dummy]
    CreateDummyRequest:
      type: object
      required: [username, full_name]
      properties:
        username:
          type: string
        full_name:
          type: string
        display_name:
          type: string
    UpdateDummyRequest:
      type: object
      properties:
        full_name:
          type: string
        display_name:
          type: string
          nullable: true